+++ apps/web/lib/api/groups/upsert-group-move-rules.ts
@@ -1,7 +1,7 @@
+import type { GroupMoveRules } from "@/lib/api/workflows/move-group/schema";
 import { getPlanCapabilities } from "@/lib/plan-capabilities";
 import { prisma } from "@/lib/prisma";
 import { WorkflowAction, WorkspaceProps } from "@/lib/types";
-import type { GroupMoveRules } from "@/lib/zod/schemas/group-move-workflows";
 import { WORKFLOW_ACTION_TYPES } from "@/lib/zod/schemas/workflows";
 import { pluralize } from "@dub/utils";
 import { PartnerGroup, WorkflowTrigger } from "@prisma/client";
