+++ apps/web/lib/zod/schemas/groups.ts
@@ -5,9 +5,9 @@ import { validSlugRegex } from "@dub/utils";
 import { PartnerLinkStructure } from "@prisma/client";
 import slugify from "@sindresorhus/slugify";
 import * as z from "zod/v4";
+import { groupMoveRulesSchema } from "../../api/workflows/move-group/schema";
 import { DiscountSchema } from "./discount";
 import { GroupBountySummarySchema } from "./group-bounties";
-import { groupMoveRulesSchema } from "./group-move-workflows";
 import { booleanQuerySchema, getPaginationQuerySchema } from "./misc";
 import { programApplicationFormSchema } from "./program-application-form";
 import { programLanderSchema } from "./program-lander";
